refactor(bff): embed bff libs within `api` library type
