--- libs/blog-bff/articles/README.md
@@ -1,7 +1,0 @@
-# api
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test api` to execute the unit tests via [Jest](https://jestjs.io).
