+++ libs/blog-bff/articles/api/README.md
@@ -1,0 +1,7 @@
+# api-articles
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test api-articles` to execute the unit tests via [Jest](https://jestjs.io).
