+++ libs/blog-bff/articles/api/jest.config.ts
@@ -1,11 +1,11 @@
 /* eslint-disable */
 export default {
-  displayName: 'api',
-  preset: '../../../jest.preset.js',
+  displayName: 'api-articles',
+  preset: '../../../../jest.preset.js',
   testEnvironment: 'node',
   transform: {
     '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../../coverage/libs/blog-bff/articles',
+  coverageDirectory: '../../../../coverage/libs/blog-bff/articles/api',
 };
