+++ libs/blog-bff/authors/api/README.md
@@ -1,7 +1,7 @@
-# blog-bff-authors
+# api-authors
 
 This library was generated with [Nx](https://nx.dev).
 
 ## Running unit tests
 
-Run `nx test blog-bff-authors` to execute the unit tests.
+Run `nx test api-authors` to execute the unit tests.
